Some Render refactoring

+++ src/core/game/graphics/GameRenderer.cpp
@@ -46,7 +46,7 @@
 #include "game/logic/GameInputManager.h"
 #include "framework/network/server/NetworkManagerServer.h"
 #include "framework/network/portable/MachineAddress.h"
-#include "framework/math/MathUtil.h"
+#include "framework/util/MathUtil.h"
 #include "framework/audio/portable/NGAudioEngine.h"
 #include "game/logic/Server.h"
 #include "game/logic/TitleEngine.h"
@@ -166,12 +166,12 @@ void GameRenderer::createDeviceDependentResources()
     _rendererHelper->createDeviceDependentResources();
     _textureManager->createDeviceDependentResources();
     
-    _parallaxLayer0FactorX = NG_CFG->getFloat("ParallaxLayer0FactorX");
-    _parallaxLayer0FactorY = NG_CFG->getFloat("ParallaxLayer0FactorY");
-    _parallaxLayer1FactorX = NG_CFG->getFloat("ParallaxLayer1FactorX");
-    _parallaxLayer1FactorY = NG_CFG->getFloat("ParallaxLayer1FactorY");
-    _parallaxLayer2FactorX = NG_CFG->getFloat("ParallaxLayer2FactorX");
-    _parallaxLayer2FactorY = NG_CFG->getFloat("ParallaxLayer2FactorY");
+    _parallaxLayer0FactorX = NG_CFG->getDouble("ParallaxLayer0FactorX");
+    _parallaxLayer0FactorY = NG_CFG->getDouble("ParallaxLayer0FactorY");
+    _parallaxLayer1FactorX = NG_CFG->getDouble("ParallaxLayer1FactorX");
+    _parallaxLayer1FactorY = NG_CFG->getDouble("ParallaxLayer1FactorY");
+    _parallaxLayer2FactorX = NG_CFG->getDouble("ParallaxLayer2FactorX");
+    _parallaxLayer2FactorY = NG_CFG->getDouble("ParallaxLayer2FactorY");
     _playerLightColor[0] = NG_CFG->getFloat("PlayerLightColorR");
     _playerLightColor[1] = NG_CFG->getFloat("PlayerLightColorG");
     _playerLightColor[2] = NG_CFG->getFloat("PlayerLightColorB");
@@ -289,6 +289,7 @@ void GameRenderer::updateCamera()
     }
     
     _camBounds[3]->getLowerLeft().set(x, y);
+    
     _camBounds[2]->getLowerLeft().set(x * _parallaxLayer2FactorX, y * _parallaxLayer2FactorY);
     _camBounds[1]->getLowerLeft().set(x * _parallaxLayer1FactorX, y * _parallaxLayer1FactorY);
     _camBounds[0]->getLowerLeft().set(x * _parallaxLayer0FactorX, y * _parallaxLayer0FactorY);
